Daos,Interfaces y Modelos de vistas
Se agregaron los daos para Combos y para tipos de combos
Así mismo se agregaron modelos de las diferentes vistas creadas y las interfaces respectivas a los daos creados

+++ src/main/java/modelo/Combos.java
@@ -6,7 +6,7 @@ public class Combos {
     
     private int combo_id;
     private String combonombre;
-    private byte combodescuento;
+    private double combodescuento;
     private byte tipo_combo_id;
 
     public Combos() {
@@ -35,11 +35,11 @@ public class Combos {
         this.combonombre = combonombre;
     }
 
-    public byte getCombodescuento() {
+    public double getCombodescuento() {
         return combodescuento;
     }
 
-    public void setCombodescuento(byte combodescuento) {
+    public void setCombodescuento(double combodescuento) {
         this.combodescuento = combodescuento;
     }
